+++ Crashlytics/Crashlytics/Components/FIRCLSProcess.c
@@ -167,7 +167,7 @@ static bool FIRCLSProcessGetThreadState(FIRCLSProcess *process,
                                         thread_t thread,
                                         FIRCLSThreadContext *context) {
   if (!FIRCLSIsValidPointer(context)) {
-    FIRCLSSDKLogError("invalid context supplied");
+    FIRCLSSDKLogError("Invalid context supplied\n");
     return false;
   }
 
@@ -259,7 +259,7 @@ static const char *FIRCLSProcessGetThreadDispatchQueueName(FIRCLSProcess *proces
   infoCount = THREAD_IDENTIFIER_INFO_COUNT;
   if (thread_info(thread, THREAD_IDENTIFIER_INFO, (thread_info_t)&info, &infoCount) !=
       KERN_SUCCESS) {
-    FIRCLSSDKLog("unable to get thread info\n");
+    FIRCLSSDKLog("Unable to get thread info\n");
     return NULL;
   }
 
@@ -395,12 +395,12 @@ static bool FIRCLSProcessRecordThread(FIRCLSProcess *process, thread_t thread, F
   FIRCLSThreadContext context;
 
   if (!FIRCLSProcessGetThreadState(process, thread, &context)) {
-    FIRCLSSDKLogError("unable to get thread state");
+    FIRCLSSDKLogError("Unable to get thread state\n");
     return false;
   }
 
   if (!FIRCLSUnwindInit(&unwindContext, context)) {
-    FIRCLSSDKLog("unable to init unwind context\n");
+    FIRCLSSDKLog("Unable to init unwind context\n");
 
     return false;
   }
@@ -490,7 +490,7 @@ bool FIRCLSProcessRecordAllThreads(FIRCLSProcess *process, FIRCLSFile *file) {
 
     FIRCLSSDKLogInfo("recording thread %d data\n", i);
     if (!FIRCLSProcessRecordThread(process, thread, file)) {
-      FIRCLSSDKLogError("Failed to record thread state. Closing threads JSON to prevent malformed crash report.");
+      FIRCLSSDKLogError("Failed to record thread state. Closing threads JSON to prevent malformed crash report.\n");
 
       FIRCLSFileWriteArrayEnd(file);
 
@@ -503,7 +503,7 @@ bool FIRCLSProcessRecordAllThreads(FIRCLSProcess *process, FIRCLSFile *file) {
 
   FIRCLSFileWriteSectionEnd(file);
 
-  FIRCLSSDKLogInfo("completed recording all thread data\n");
+  FIRCLSSDKLogInfo("Completed recording all thread data\n");
 
   return true;
 }