+++ Crashlytics/CHANGELOG.md
@@ -1,3 +1,13 @@
+# v7.8.0
+- [added] Added a new API checkAndUpdateUnsentReportsWithCompletion for updating the crash report from the previous run of the app if, for example, the developer wants to implement a feedback dialog to ask end-users for more information. Unsent Crashlytics Reports have familiar methods like setting custom keys and logs (#7503).
+- [changed] Added a limit to the number of unsent reports on disk to prevent disk filling up when automatic data collection is off. Developers can ensure this limit is never reached by calling send/deleteUnsentReports every run (#7619).
+
+# v7.7.0
+- [added] Added a new API to allow for bulk logging of custom keys and values (#7302).
+
+# v7.6.0
+- [fixed] Fixed an issue where some developers experienced a race condition involving binary image operations (#7459).
+
 # v7.5.0
 - [changed] Improve start-up performance by moving some initialization work to a background thread (#7332).
 - [changed] Updated upload-symbols to a version that is notarized to avoid macOS security alerts (#7323).